Make test utils available outside of clock_nanosleep test
